+++ ads/model/model_metadata.py
@@ -31,7 +31,6 @@ try:
 except:
     from yaml import Dumper as dumper
 
-
 logging.basicConfig(stream=sys.stdout, level=logging.INFO)
 logger = logging.getLogger("ADS")
 
@@ -256,9 +255,9 @@ class ModelMetadataItem(ABC):
         return json.dumps(self.to_dict())
 
     def to_json_file(
-        self,
-        file_path: str,
-        storage_options: dict = None,
+            self,
+            file_path: str,
+            storage_options: dict = None,
     ) -> None:
         """Saves the metadata item value to a local file or object storage.
 
@@ -313,9 +312,9 @@ class ModelMetadataItem(ABC):
             storage_options = factory.default_storage_options or {"config": {}}
 
         with fsspec.open(
-            file_path,
-            mode="w",
-            **(storage_options),
+                file_path,
+                mode="w",
+                **(storage_options),
         ) as f:
             f.write(json.dumps(self.value))
 
@@ -400,9 +399,9 @@ class ModelTaxonomyMetadataItem(ModelMetadataItem):
     _FIELDS = ["key", "value"]
 
     def __init__(
-        self,
-        key: str,
-        value: str = None,
+            self,
+            key: str,
+            value: str = None,
     ):
         self.key = key
         self.value = value
@@ -500,17 +499,17 @@ class ModelTaxonomyMetadataItem(ModelMetadataItem):
             If invalid Framework provided.
         """
         if (
-            self.key.lower() == MetadataTaxonomyKeys.USE_CASE_TYPE.lower()
-            and self.value
-            and (not isinstance(self.value, str) or self.value not in UseCaseType)
+                self.key.lower() == MetadataTaxonomyKeys.USE_CASE_TYPE.lower()
+                and self.value
+                and (not isinstance(self.value, str) or self.value not in UseCaseType)
         ):
             raise ValueError(
                 f"Invalid value of `UseCaseType`. Choose from {UseCaseType.values()}."
             )
         if (
-            self.key.lower() == MetadataTaxonomyKeys.FRAMEWORK.lower()
-            and self.value
-            and (not isinstance(self.value, str) or self.value not in Framework)
+                self.key.lower() == MetadataTaxonomyKeys.FRAMEWORK.lower()
+                and self.value
+                and (not isinstance(self.value, str) or self.value not in Framework)
         ):
             raise ValueError(
                 f"Invalid value of `Framework`. Choose from {Framework.values()}."
@@ -557,11 +556,11 @@ class ModelCustomMetadataItem(ModelTaxonomyMetadataItem):
     _FIELDS = ["key", "value", "description", "category"]
 
     def __init__(
-        self,
-        key: str,
-        value: str = None,
-        description: str = None,
-        category: str = None,
+            self,
+            key: str,
+            value: str = None,
+            description: str = None,
+            category: str = None,
     ):
         super().__init__(key=key, value=value)
         self.description = description
@@ -689,8 +688,8 @@ class ModelCustomMetadataItem(ModelTaxonomyMetadataItem):
                 raise MetadataValueTooLong(self.key, len(value))
 
         if (
-            self.description
-            and len(self.description) > METADATA_DESCRIPTION_LENGTH_LIMIT
+                self.description
+                and len(self.description) > METADATA_DESCRIPTION_LENGTH_LIMIT
         ):
             raise MetadataDescriptionTooLong(self.key, len(self.description))
 
@@ -730,7 +729,7 @@ class ModelMetadata(ABC):
         self._items = set()
 
     def get(
-        self, key: str, value: Optional[Any] = _sentinel
+            self, key: str, value: Optional[Any] = _sentinel
     ) -> Union[ModelMetadataItem, Any]:
         """Returns the model metadata item by provided key.
 
@@ -891,9 +890,9 @@ class ModelMetadata(ABC):
         return [item._to_oci_metadata() for item in self._items]
 
     def to_json_file(
-        self,
-        file_path: str,
-        storage_options: dict = None,
+            self,
+            file_path: str,
+            storage_options: dict = None,
     ) -> None:
         """Saves the metadata to a local file or object storage.
 
@@ -948,9 +947,9 @@ class ModelMetadata(ABC):
             storage_options = factory.default_storage_options or {"config": {}}
 
         with fsspec.open(
-            file_path,
-            mode="w",
-            **(storage_options),
+                file_path,
+                mode="w",
+                **(storage_options),
         ) as f:
             f.write(self.to_json())
 
@@ -1082,12 +1081,12 @@ class ModelCustomMetadata(ModelMetadata):
         self._items = set()
 
     def add(
-        self,
-        key: str,
-        value: str,
-        description: str = "",
-        category: str = MetadataCustomCategory.OTHER,
-        replace: bool = False,
+            self,
+            key: str,
+            value: str,
+            description: str = "",
+            category: str = MetadataCustomCategory.OTHER,
+            replace: bool = False,
     ) -> None:
         """Adds a new model metadata item. Overrides the existing one if replace flag is True.
 
@@ -1396,10 +1395,10 @@ class ModelCustomMetadata(ModelMetadata):
             In case of the wrong input data format.
         """
         if (
-            not data
-            or not isinstance(data, Dict)
-            or not "data" in data
-            or not isinstance(data["data"], List)
+                not data
+                or not isinstance(data, Dict)
+                or not "data" in data
+                or not isinstance(data["data"], List)
         ):
             raise ValueError(
                 "An error occurred when attempting to deserialize the model custom metadata from a dictionary. "
@@ -1548,10 +1547,10 @@ class ModelTaxonomyMetadata(ModelMetadata):
             In case of the wrong input data format.
         """
         if (
-            not data
-            or not isinstance(data, Dict)
-            or not "data" in data
-            or not isinstance(data["data"], List)
+                not data
+                or not isinstance(data, Dict)
+                or not "data" in data
+                or not isinstance(data["data"], List)
         ):
             raise ValueError(
                 "An error occurred when attempting to deserialize the model taxonomy metadata from a dictionary. "
@@ -1586,10 +1585,10 @@ class ModelProvenanceMetadata(DataClassSerializable):
 
     @classmethod
     def fetch_training_code_details(
-        cls,
-        training_script_path: str = None,
-        training_id: str = None,
-        artifact_dir: str = None,
+            cls,
+            training_script_path: str = None,
+            training_id: str = None,
+            artifact_dir: str = None,
     ):
         """Fetches the training code details: repo, git_branch, git_commit, repository_url, training_script_path and training_id.
 
@@ -1672,15 +1671,15 @@ class ModelProvenanceMetadata(DataClassSerializable):
         if self.repo is not None and not ignore:
             path_abs = os.path.abspath(path)
             if (
-                os.path.commonpath([path_abs, self.repo.working_dir])
-                == self.repo.working_dir
+                    os.path.commonpath([path_abs, self.repo.working_dir])
+                    == self.repo.working_dir
             ):
                 path_relpath = os.path.relpath(path_abs, self.repo.working_dir)
                 if self.repo.is_dirty(path=path_relpath) or any(
-                    [
-                        os.path.commonpath([path_relpath, untracked]) == path_relpath
-                        for untracked in self.repo.untracked_files
-                    ]
+                        [
+                            os.path.commonpath([path_relpath, untracked]) == path_relpath
+                            for untracked in self.repo.untracked_files
+                        ]
                 ):
                     raise ChangesNotCommitted(path_abs)
 
@@ -1703,7 +1702,7 @@ class ModelProvenanceMetadata(DataClassSerializable):
 
     @classmethod
     def _from_oci_metadata(
-        cls, model_provenance: oci.data_science.models.ModelProvenance
+            cls, model_provenance: oci.data_science.models.ModelProvenance
     ) -> "ModelProvenanceMetadata":
         """Creates a new model provenance metadata item from the `oci.data_science.models.ModelProvenance` object.
 
@@ -1760,392 +1759,4 @@ class ModelProvenanceMetadata(DataClassSerializable):
         string
             Serialized version of object as a YAML string
         """
-        return self.to_yaml()
-
-
-class CustomerNotificationType(str, metaclass=ExtendedEnumMeta):
-    NONE = "NONE"
-    ALL = "ALL"
-    ON_FAILURE = "ON_FAILURE"
-    ON_SUCCESS = "ON_SUCCESS"
-
-    _value_map = {
-        "NONE": NONE,
-        "ALL": ALL,
-        "ON_FAILURE": ON_FAILURE,
-        "ON_SUCCESS": ON_SUCCESS,
-    }
-
-    @classmethod
-    def create(cls, key):
-        if key in cls._value_map:
-            return cls._value_map[key]
-        raise ValueError(f"Invalid CustomerNotificationType: {key}")
-
-    @classmethod
-    def from_string(cls, value):
-        for member in cls:
-            if member.value == value:
-                return member
-        raise ValueError(f"Invalid CustomerNotificationType: {value}")
-
-    @classmethod
-    def is_valid(cls, value):
-        return value in (cls.NONE, cls.ALL, cls.ON_FAILURE, cls.ON_SUCCESS)
-
-    @property
-    def value(self):
-        return str(self)
-
-
-class ModelBackupSetting:
-    """
-    Class that represents Model Backup Setting Details Metadata.
-
-    Methods
-    -------
-    to_dict(self) -> Dict:
-        Serializes the backup settings into a dictionary.
-    from_dict(cls, data: Dict) -> 'ModelBackupSetting':
-        Constructs backup settings from a dictionary.
-    to_json(self) -> str:
-        Serializes the backup settings into a JSON string.
-    from_json(cls, json_str: str) -> 'ModelBackupSetting':
-        Constructs backup settings from a JSON string.
-    to_yaml(self) -> str:
-        Serializes the backup settings into a YAML string.
-    validate(self) -> bool:
-        Validates the backup settings details.
-    """
-
-    def __init__(
-        self,
-        is_backup_enabled: Optional[bool] = None,
-        backup_region: Optional[str] = None,
-        customer_notification_type: Optional[CustomerNotificationType] = None,
-    ):
-        self.is_backup_enabled = (
-            is_backup_enabled if is_backup_enabled is not None else False
-        )
-        self.backup_region = backup_region
-        self.customer_notification_type = (
-            customer_notification_type or CustomerNotificationType.NONE
-        )
-
-    def to_dict(self) -> Dict:
-        """Serializes the backup settings into a dictionary."""
-        return {
-            "is_backup_enabled": self.is_backup_enabled,
-            "backup_region": self.backup_region,
-            "customer_notification_type": self.customer_notification_type,
-        }
-
-    @classmethod
-    def from_dict(cls, data: Dict) -> "ModelBackupSetting":
-        """Constructs backup settings from a dictionary."""
-        return cls(
-            is_backup_enabled=data.get("is_backup_enabled"),
-            backup_region=data.get("backup_region"),
-            customer_notification_type=CustomerNotificationType(
-                data.get("customer_notification_type")
-            )
-            or None,
-        )
-
-    def to_json(self) -> str:
-        """Serializes the backup settings into a JSON string."""
-        return json.dumps(self.to_dict())
-
-    @classmethod
-    def from_json(cls, json_str) -> "ModelBackupSetting":
-        """Constructs backup settings from a JSON string or dictionary."""
-        if isinstance(json_str, str):
-            data = json.loads(json_str)
-        else:
-            data = json_str  # Assume it's already a dictionary or appropriate type
-
-        return cls.from_dict(data)
-
-    def to_yaml(self) -> str:
-        """Serializes the backup settings into a YAML string."""
-        return yaml.dump(self.to_dict())
-
-    def validate(self) -> bool:
-        """Validates the backup settings details. Returns True if valid, False otherwise."""
-        if not isinstance(self.is_backup_enabled, bool):
-            return False
-        if self.backup_region and not isinstance(self.backup_region, str):
-            return False
-        if not isinstance(self.customer_notification_type, str) \
-                or not CustomerNotificationType.is_valid(self.customer_notification_type):
-            return False
-        return True
-
-    def __repr__(self):
-        return self.to_yaml()
-
-
-class ModelRetentionSetting:
-    """
-    Class that represents Model Retention Setting Details Metadata.
-
-    Methods
-    -------
-    to_dict(self) -> Dict:
-        Serializes the retention settings into a dictionary.
-    from_dict(cls, data: Dict) -> 'ModelRetentionSetting':
-        Constructs retention settings from a dictionary.
-    to_json(self) -> str:
-        Serializes the retention settings into a JSON string.
-    from_json(cls, json_str: str) -> 'ModelRetentionSetting':
-        Constructs retention settings from a JSON string.
-    to_yaml(self) -> str:
-        Serializes the retention settings into a YAML string.
-    validate(self) -> bool:
-        Validates the retention settings details.
-    """
-
-    def __init__(
-        self,
-        archive_after_days: Optional[int] = None,
-        delete_after_days: Optional[int] = None,
-        customer_notification_type: Optional[CustomerNotificationType] = None,
-    ):
-        self.archive_after_days = archive_after_days
-        self.delete_after_days = delete_after_days
-        self.customer_notification_type = (
-            customer_notification_type or CustomerNotificationType.NONE
-        )
-
-    def to_dict(self) -> Dict:
-        """Serializes the retention settings into a dictionary."""
-        return {
-            "archive_after_days": self.archive_after_days,
-            "delete_after_days": self.delete_after_days,
-            "customer_notification_type": self.customer_notification_type,
-        }
-
-    @classmethod
-    def from_dict(cls, data: Dict) -> "ModelRetentionSetting":
-        """Constructs retention settings from a dictionary."""
-        return cls(
-            archive_after_days=data.get("archive_after_days"),
-            delete_after_days=data.get("delete_after_days"),
-            customer_notification_type=CustomerNotificationType(
-                data.get("customer_notification_type")
-            )
-            or None,
-        )
-
-    def to_json(self) -> str:
-        """Serializes the retention settings into a JSON string."""
-        return json.dumps(self.to_dict())
-
-    @classmethod
-    def from_json(cls, json_str) -> "ModelRetentionSetting":
-        """Constructs retention settings from a JSON string."""
-        if isinstance(json_str, str):
-            data = json.loads(json_str)
-        else:
-            data = json_str
-        return cls.from_dict(data)
-
-    def to_yaml(self) -> str:
-        """Serializes the retention settings into a YAML string."""
-        return yaml.dump(self.to_dict())
-
-    def validate(self) -> bool:
-        """Validates the retention settings details. Returns True if valid, False otherwise."""
-        if self.archive_after_days is not None and (
-            not isinstance(self.archive_after_days, int) or self.archive_after_days < 0
-        ):
-            return False
-        if self.delete_after_days is not None and (
-            not isinstance(self.delete_after_days, int) or self.delete_after_days < 0
-        ):
-            return False
-        if not isinstance(self.customer_notification_type, str) or not \
-                CustomerNotificationType.is_valid(self.customer_notification_type):
-            return False
-        return True
-
-    def __repr__(self):
-        return self.to_yaml()
-
-
-class SettingStatus(str, metaclass=ExtendedEnumMeta):
-    """Enum to represent the status of retention settings."""
-
-    PENDING = "PENDING"
-    SUCCEEDED = "SUCCEEDED"
-    FAILED = "FAILED"
-
-    @classmethod
-    def is_valid(cls, state: str) -> bool:
-        """Validates the given state against allowed SettingStatus values."""
-        return state in (cls.PENDING, cls.SUCCEEDED, cls.FAILED)
-
-
-class ModelRetentionOperationDetails:
-    """
-    Class that represents Model Retention Operation Details Metadata.
-
-    Methods
-    -------
-    to_dict(self) -> Dict:
-        Serializes the retention operation details into a dictionary.
-    from_dict(cls, data: Dict) -> 'ModelRetentionOperationDetails':
-        Constructs retention operation details from a dictionary.
-    to_json(self) -> str:
-        Serializes the retention operation details into a JSON string.
-    from_json(cls, json_str: str) -> 'ModelRetentionOperationDetails':
-        Constructs retention operation details from a JSON string.
-    to_yaml(self) -> str:
-        Serializes the retention operation details into a YAML string.
-    validate(self) -> bool:
-        Validates the retention operation details.
-    """
-
-    def __init__(
-        self,
-        archive_state: Optional[SettingStatus] = None,
-        archive_state_details: Optional[str] = None,
-        delete_state: Optional[SettingStatus] = None,
-        delete_state_details: Optional[str] = None,
-        time_archival_scheduled: Optional[int] = None,
-        time_deletion_scheduled: Optional[int] = None,
-    ):
-        self.archive_state = archive_state
-        self.archive_state_details = archive_state_details
-        self.delete_state = delete_state
-        self.delete_state_details = delete_state_details
-        self.time_archival_scheduled = time_archival_scheduled
-        self.time_deletion_scheduled = time_deletion_scheduled
-
-    def to_dict(self) -> Dict:
-        """Serializes the retention operation details into a dictionary."""
-        return {
-            "archive_state": self.archive_state or None,
-            "archive_state_details": self.archive_state_details,
-            "delete_state": self.delete_state or None,
-            "delete_state_details": self.delete_state_details,
-            "time_archival_scheduled": self.time_archival_scheduled,
-            "time_deletion_scheduled": self.time_deletion_scheduled,
-        }
-
-    @classmethod
-    def from_dict(cls, data: Dict) -> "ModelRetentionOperationDetails":
-        """Constructs retention operation details from a dictionary."""
-        return cls(
-            archive_state=SettingStatus(data.get("archive_state")) or None,
-            archive_state_details=data.get("archive_state_details"),
-            delete_state=SettingStatus(data.get("delete_state")) or None,
-            delete_state_details=data.get("delete_state_details"),
-            time_archival_scheduled=data.get("time_archival_scheduled"),
-            time_deletion_scheduled=data.get("time_deletion_scheduled"),
-        )
-
-    def to_json(self) -> str:
-        """Serializes the retention operation details into a JSON string."""
-        return json.dumps(self.to_dict())
-
-    @classmethod
-    def from_json(cls, json_str: str) -> "ModelRetentionOperationDetails":
-        """Constructs retention operation details from a JSON string."""
-        data = json.loads(json_str)
-        return cls.from_dict(data)
-
-    def to_yaml(self) -> str:
-        """Serializes the retention operation details into a YAML string."""
-        return yaml.dump(self.to_dict())
-
-    def validate(self) -> bool:
-        """Validates the retention operation details."""
-        return all(
-            [
-                self.archive_state is None or SettingStatus.is_valid(self.archive_state),
-                self.delete_state is None or SettingStatus.is_valid(self.delete_state),
-                self.time_archival_scheduled is None
-                or isinstance(self.time_archival_scheduled, int),
-                self.time_deletion_scheduled is None
-                or isinstance(self.time_deletion_scheduled, int),
-            ]
-        )
-
-    def __repr__(self):
-        return self.to_yaml()
-
-
-class ModelBackupOperationDetails:
-    """
-    Class that represents Model Backup Operation Details Metadata.
-
-    Methods
-    -------
-    to_dict(self) -> Dict:
-        Serializes the backup operation details into a dictionary.
-    from_dict(cls, data: Dict) -> 'ModelBackupOperationDetails':
-        Constructs backup operation details from a dictionary.
-    to_json(self) -> str:
-        Serializes the backup operation details into a JSON string.
-    from_json(cls, json_str: str) -> 'ModelBackupOperationDetails':
-        Constructs backup operation details from a JSON string.
-    to_yaml(self) -> str:
-        Serializes the backup operation details into a YAML string.
-    validate(self) -> bool:
-        Validates the backup operation details.
-    """
-
-    def __init__(
-        self,
-        backup_state: Optional[SettingStatus] = None,
-        backup_state_details: Optional[str] = None,
-        time_last_backed_up: Optional[int] = None,
-    ):
-        self.backup_state = backup_state
-        self.backup_state_details = backup_state_details
-        self.time_last_backed_up = time_last_backed_up
-
-    def to_dict(self) -> Dict:
-        """Serializes the backup operation details into a dictionary."""
-        return {
-            "backup_state": self.backup_state or None,
-            "backup_state_details": self.backup_state_details,
-            "time_last_backed_up": self.time_last_backed_up,
-        }
-
-    @classmethod
-    def from_dict(cls, data: Dict) -> "ModelBackupOperationDetails":
-        """Constructs backup operation details from a dictionary."""
-        return cls(
-            backup_state=SettingStatus(data.get("backup_state")) or None,
-            backup_state_details=data.get("backup_state_details"),
-            time_last_backed_up=data.get("time_last_backed_up"),
-        )
-
-    def to_json(self) -> str:
-        """Serializes the backup operation details into a JSON string."""
-        return json.dumps(self.to_dict())
-
-    @classmethod
-    def from_json(cls, json_str: str) -> "ModelBackupOperationDetails":
-        """Constructs backup operation details from a JSON string."""
-        data = json.loads(json_str)
-        return cls.from_dict(data)
-
-    def to_yaml(self) -> str:
-        """Serializes the backup operation details into a YAML string."""
-        return yaml.dump(self.to_dict())
-
-    def validate(self) -> bool:
-        """Validates the backup operation details."""
-        if self.backup_state is not None and not SettingStatus.is_valid(self.backup_state):
-            return False
-        if self.time_last_backed_up is not None and not isinstance(
-            self.time_last_backed_up, int
-        ):
-            return False
-        return True
-
-    def __repr__(self):
         return self.to_yaml()
